refactor(mobile): extract footnote utilities to shared module and fix hardcoded color
- Create mobile/src/utils/footnoteUtils.ts with sanitizeEmTags, escapeRegex,
  toSuperscriptNumber, createFootnoteLookup, and buildMarkerRegex
- Remove duplicate definitions from VerseCard.tsx and VerseDetailContent.tsx
- Add markerColor parameter to renderTranslationSegment/renderTranslationFlowText
  in VerseDetailContent.tsx and pass colors.accentCopper from the call site
  instead of hardcoded #B4834D

Agent-Logs-Url: https://github.com/edyhvh/davar/sessions/cdfb8ec3-6926-465a-ae09-a00ad4e4dcb9
